refactor(graphql): move graphql to server/graphql
