+++ server/graphql/models/Error.js
