+++ server/graphql/models/Installation.js
